pyproject.toml
[tool.poetry]
name = "e-commerce-api"
version = "0.1.0"
description = ""
authors = ["kazair <tokubd127@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
tortoise-orm = "^0.21.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.9.2"}
bcrypt = "4.0.1"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.1"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
